// frontend/src/routes/oc-declare/types/OCDeclareArcLabel.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ObjectTypeAssociation } from "./ObjectTypeAssociation";

export type OCDeclareArcLabel = {
	each: Array<ObjectTypeAssociation>;
	any: Array<ObjectTypeAssociation>;
	all: Array<ObjectTypeAssociation>;
};